Removed more unused graph exporters
